# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_UserRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "totalPages": {
            "title": "Totalpages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "title": "Page[UserRead]",
        "type": "object"
      },
      "PasswordReset": {
        "description": "The response to resetting a password.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "temporaryPassword": {
            "description": "Shown once.",
            "title": "Temporarypassword",
            "type": "string"
          }
        },
        "required": [
          "temporaryPassword",
          "message"
        ],
        "title": "PasswordReset",
        "type": "object"
      },
...
      "RoleRead": {
        "description": "A role with the permissions attached to it.\n\n``permissions`` is served from here rather than hard-coded in the frontend so the\nRoles screen is data-driven — a table that has to be edited in two places to stay\ntrue will eventually be true in only one.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "roleId": {
            "title": "Roleid",
            "type": "integer"
          },
          "userCount": {
            "default": 0,
            "title": "Usercount",
            "type": "integer"
          }
        },
        "required": [
          "roleId",
          "name",
          "displayName"
        ],
        "title": "RoleRead",
        "type": "object"
      },
...
      "UserCreateInput": {
        "description": "Creating a user (FR-12).\n\nNo password field: the server generates a temporary one. Letting an administrator\nchoose another person's initial password means it travels by whatever channel they\nuse to pass it on and is very often reused.",
        "properties": {
          "contactNumber": {
            "anyOf": [
              {
                "maxLength": 24,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required when the role is TRAINER. A trainer who cannot be reached cannot be offered an assignment.",
            "title": "Contactnumber"
          },
          "directorateId": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required when the role is TRAINER.",
            "title": "Directorateid"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "forceNumber": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required when the role is TRAINER.",
            "title": "Forcenumber"
          },
          "fullName": {
            "maxLength": 150,
            "minLength": 3,
            "title": "Fullname",
            "type": "string"
          },
          "rankId": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rankid"
          },
          "role": {
            "description": "One of the four role names.",
            "title": "Role",
            "type": "string"
          },
          "stationId": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required when the role is TRAINER — a trainer must have a posting.",
            "title": "Stationid"
          },
          "username": {
            "maxLength": 60,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9._-]+$",
            "title": "Username",
            "type": "string"
          },
          "yearsExperience": {
            "anyOf": [
              {
                "maximum": 60.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yearsexperience"
          }
        },
        "required": [
          "username",
          "fullName",
          "email",
          "role"
        ],
        "title": "UserCreateInput",
        "type": "object"
      },
      "UserCreated": {
        "description": "The response to creating a user.\n\n``temporaryPassword`` is returned **once, here, in the body**. It is never logged,\nnever stored in plaintext, never emailed by this system, and never retrievable\nagain — a password that can be looked up later is not a password.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "temporaryPassword": {
            "description": "Shown once. The user must change it at first sign-in.",
            "title": "Temporarypassword",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "required": [
          "user",
          "temporaryPassword",
          "message"
        ],
        "title": "UserCreated",
        "type": "object"
      },
      "UserRead": {
        "description": "A user account. **Never carries a password field, in any form.**",
        "properties": {
          "accountStatus": {
            "title": "Accountstatus",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "fullName": {
            "title": "Fullname",
            "type": "string"
          },
          "lastLoginAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastloginat"
          },
          "mustChangePassword": {
            "title": "Mustchangepassword",
            "type": "boolean"
          },
          "rank": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Police rank code, where the user has one.",
            "title": "Rank"
          },
          "role": {
            "description": "Machine name, e.g. TRAINING_ADMINISTRATOR.",
            "title": "Role",
            "type": "string"
          },
          "roleDisplayName": {
            "default": "",
            "title": "Roledisplayname",
            "type": "string"
          },
          "roleId": {
            "title": "Roleid",
            "type": "integer"
          },
          "trainerId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when the account is linked to a trainer profile.",
            "title": "Trainerid"
          },
          "userId": {
            "title": "Userid",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "username",
          "fullName",
          "email",
          "role",
          "roleId",
          "accountStatus",
          "mustChangePassword",
          "createdAt"
        ],
        "title": "UserRead",
        "type": "object"
      },
...
      "UserUpdateInput": {
        "description": "Editing a user (FR-12).",
        "properties": {
          "accountStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ACTIVE or SUSPENDED.",
            "title": "Accountstatus"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "fullName": {
            "anyOf": [
              {
                "maxLength": 150,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fullname"
          },
          "rankId": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rankid"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "title": "UserUpdateInput",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT access token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "ict@upf.go.ug",
      "name": "Uganda Police Force — Directorate of Human Resource Development"
    },
...
    "license": {
      "name": "Restricted — Uganda Police Force internal use"
    },
    "title": "Trainer Prediction System API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/roles": {
      "get": {
        "description": "Serves the permission matrix as **data**, so the Roles screen is not a hard-coded table in the frontend that has to be edited in two places to stay true — and would therefore eventually be true in one.",
        "operationId": "list_roles_api_v1_roles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleRead"
                  },
                  "title": "Response List Roles Api V1 Roles Get",
                  "type": "array"
                }
              }
            },
            "description": "The roles, with user counts and permissions."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The four roles and what each may do",
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "description": "Filterable by role, account status, and free text over username, name and email.\n\nReturned in the standard paginated envelope. There are forty-odd accounts today, but *every* list endpoint uses the envelope — an unpaginated list is a latent problem waiting for the system to succeed.",
        "operationId": "list_users_api_v1_users_get",
        "parameters": [
          {
            "description": "Username, name, or email.",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Username, name, or email.",
              "title": "Search"
            }
          },
          {
            "description": "Role name.",
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Role name.",
              "title": "Role"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sortby"
            }
          },
          {
            "in": "query",
            "name": "sortDir",
            "required": false,
            "schema": {
              "default": "asc",
              "pattern": "^(asc|desc)$",
              "title": "Sortdir",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UserRead_"
                }
              }
            },
            "description": "A page of users."
          },
          "403": {
            "description": "System Administrator only."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "Creates the account with a **generated** temporary password and `mustChangePassword = true`.\n\nThe caller does not choose the password. An initial password chosen by an administrator travels by whatever channel they use to pass it on, and is very often reused.\n\nIf the role is `TRAINER`, the linked trainer profile is created **in the same transaction** — a Trainer account without a profile cannot see assignments and cannot be ranked, which is a broken state that must be impossible rather than merely unlikely. `stationId` and `forceNumber` are therefore required for that role.\n\n`temporaryPassword` is in the response body **once**. It is never logged, never stored in plaintext, and cannot be retrieved again.",
        "operationId": "create_user_api_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreated"
                }
              }
            },
            "description": "Created. The temporary password is shown once."
          },
          "403": {
            "description": "System Administrator only."
          },
          "409": {
            "description": "The username or email is already in use."
          },
          "422": {
            "description": "A trainer account was submitted without a posting."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a user (FR-12)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "description": "Return one user.",
        "operationId": "get_user_api_v1_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "The user."
          },
          "404": {
            "description": "No such user."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "One user",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "A **role change takes effect at the next sign-in** — the role travels in the access token. Rather than leave that as a surprise, the response says so, and the user's refresh sessions are revoked so the change lands within the access token's fifteen-minute lifetime rather than at its leisure.\n\n`accountStatus` accepts `ACTIVE` or `SUSPENDED`. Deactivation has its own endpoint because it must also revoke sessions.",
        "operationId": "update_user_api_v1_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Updated."
          },
          "403": {
            "description": "System Administrator only."
          },
          "409": {
            "description": "The email is taken, deactivation was attempted here, or this is the last active System Administrator."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Edit a user (FR-12)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{user_id}/deactivate": {
      "post": {
        "description": "Sets `DEACTIVATED` and **revokes every session immediately**. A deactivated user cannot sign in with the correct password, and cannot continue on an access token issued a minute ago — the account status is re-read from the database on every request precisely so revocation is immediate.\n\n**The last active System Administrator cannot be deactivated** (409), and neither can your own account.",
        "operationId": "deactivate_user_api_v1_users__user_id__deactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Deactivated, sessions revoked."
          },
          "409": {
            "description": "Already deactivated, your own account, or the last administrator."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Deactivate a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{user_id}/reset-password": {
      "post": {
        "description": "Issues a new temporary password, sets `mustChangePassword`, clears any lockout, and revokes every session. The password is shown once.",
        "operationId": "reset_password_api_v1_users__user_id__reset_password_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordReset"
                }
              }
            },
            "description": "Reset. The temporary password is shown once."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset a user's password",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
...
}